% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndiaPIN-package.R
\docType{package}
\name{IndiaPIN-package}
\alias{IndiaPIN-package}
\alias{_PACKAGE}
\alias{testduplicate-package}
\title{IndiaPIN: Indian PIN Codes Directory}
\description{
This dataset contains the directory of all PIN codes in India with latitude and longitude details. (Updated: December 2021).
}
\author{
\strong{Maintainer}: Harshvardhan \email{hello@harsh17.in} (\href{https://orcid.org/0000-0001-8086-544X}{ORCID})

}
\keyword{internal}
